Using new Elysia instance per group and route
