+++ packages/api/src/routers/routes.ts
@@ -1,0 +1,5 @@
+import Elysia from 'elysia';
+import { authRouters } from './authRoutes';
+import { featuresRoutes } from './featuresRoutes';
+
+export const routes = new Elysia().use(authRouters).use(featuresRoutes);
